fix(zb_io): tolerate missing system CA certs when building HTTP clients
reqwests rustls feature eagerly loads the system CA trust store at
client-build time. In a packaging/build sandbox with no system certs,
Client::builder().build() fails with "No CA certificates were loaded
from the system", and the .unwrap_or_else(|_| reqwest::Client::new())
fallback re-panicked since Client::new() also panics when TLS cannot
initialize. This broke `zb run` and the cargo test suite during
sandboxed builds even though those tests only use local HTTP mocks.

Centralize a single, sandbox-tolerant rustls ClientConfig in
network::tls: prefer native roots, and fall back to the bundled
webpki-roots Mozilla roots when no system trust store is available.
Every reqwest client (ApiClient, Downloader primary + isolated, and the
source downloader) now builds via use_preconfigured_tls with that
config, and all panicking reqwest::Client::new() fallbacks are removed.

Fixes #245
