+++ Cargo.toml
@@ -16,6 +16,7 @@ indicatif = "0.18.3"
 reqwest = { version = "0.13.1", default-features = false, features = ["json", "stream", "rustls", "http2"] }
 rustls = { version = "0.23.26", features = ["aws-lc-rs"] }
 rustls-native-certs = "0.8.3"
+webpki-roots = "1.0.7"
 flate2 = "1.1.8"
 tar = "0.4.44"
 xz2 = "0.1.7"
