+++ zb_io/Cargo.toml
@@ -13,6 +13,7 @@ rayon.workspace = true
 regex.workspace = true
 rustls.workspace = true
 rustls-native-certs.workspace = true
+webpki-roots.workspace = true
 rusqlite.workspace = true
 serde.workspace = true
 serde_json.workspace = true
